feat: widen DebugCpu debug addresses u16 -> u32
- DebugCpu::debug_pc returns u32 and debug_disassemble takes addr: u32,
  per the Address Width section of docs/designs/debug-observability.md
- M68000 now reports its full 24-bit PC instead of the low half; the
  seven 16-bit CPUs widen losslessly and shim addr back to u16 for
  their Disassemble impls
- DebugDisassembly::addr widened to u32 to match
- debug_ui keeps its u16 breakpoint set and disassembly scan for now
  (widened by the 32-bit UI subtask); PC comparisons widen the stored
  u16 rather than truncating the live PC

+++ core/src/core/debug.rs
@@ -21,7 +21,7 @@ pub struct DebugRegister {
 /// Result of disassembling one instruction at a given address.
 pub struct DebugDisassembly {
     /// Address of the instruction.
-    pub addr: u16,
+    pub addr: u32,
     /// Raw instruction bytes.
     pub bytes: Vec<u8>,
     /// Formatted instruction text (e.g., "LDA  #$42").
@@ -45,13 +45,19 @@ pub trait Debuggable {
 /// and disassembly. The debugger uses this for stepping and code display.
 pub trait DebugCpu: Debuggable {
     /// Read the current program counter.
-    fn debug_pc(&self) -> u16;
+    ///
+    /// `u32` so 24/32-bit CPUs (M68000) report their full PC; 16-bit CPUs
+    /// widen losslessly.
+    fn debug_pc(&self) -> u32;
 
     /// True if the CPU is at an instruction boundary (ready to fetch next opcode).
     fn debug_at_instruction_boundary(&self) -> bool;
 
     /// Disassemble one instruction from raw bytes at the given address.
-    fn debug_disassemble(&self, addr: u16, bytes: &[u8]) -> DisassembledInstruction;
+    ///
+    /// 16-bit CPUs decode `addr` modulo their 64 KB space; their debuggers
+    /// never pass addresses above `0xFFFF`.
+    fn debug_disassemble(&self, addr: u32, bytes: &[u8]) -> DisassembledInstruction;
 }
 
 /// Debug view of a bus and its connected devices.
