+++ core/src/cpu/m6800/mod.rs
@@ -761,15 +761,15 @@ impl Debuggable for M6800 {
 }
 
 impl DebugCpu for M6800 {
-    fn debug_pc(&self) -> u16 {
-        self.pc
+    fn debug_pc(&self) -> u32 {
+        u32::from(self.pc)
     }
 
     fn debug_at_instruction_boundary(&self) -> bool {
         self.at_instruction_boundary()
     }
 
-    fn debug_disassemble(&self, addr: u16, bytes: &[u8]) -> DisassembledInstruction {
-        <Self as crate::cpu::Disassemble>::disassemble(addr, bytes)
+    fn debug_disassemble(&self, addr: u32, bytes: &[u8]) -> DisassembledInstruction {
+        <Self as crate::cpu::Disassemble>::disassemble(addr as u16, bytes)
     }
 }
