+++ core/src/cpu/z80/mod.rs
@@ -925,15 +925,15 @@ impl Debuggable for Z80 {
 }
 
 impl DebugCpu for Z80 {
-    fn debug_pc(&self) -> u16 {
-        self.pc
+    fn debug_pc(&self) -> u32 {
+        u32::from(self.pc)
     }
 
     fn debug_at_instruction_boundary(&self) -> bool {
         self.at_instruction_boundary()
     }
 
-    fn debug_disassemble(&self, addr: u16, bytes: &[u8]) -> DisassembledInstruction {
-        <Self as crate::cpu::Disassemble>::disassemble(addr, bytes)
+    fn debug_disassemble(&self, addr: u32, bytes: &[u8]) -> DisassembledInstruction {
+        <Self as crate::cpu::Disassemble>::disassemble(addr as u16, bytes)
     }
 }
